+++ packages/core/tests/schema.spec.ts
@@ -1,0 +1,34 @@
+import { GraphQLObjectType, GraphQLString, GraphQLSchema } from 'graphql';
+
+const QueryRootType = new GraphQLObjectType({
+  name: 'QueryRoot',
+  fields: {
+    test: {
+      type: GraphQLString,
+      args: {
+        who: { type: GraphQLString },
+      },
+      resolve: (_root, args: { who?: string }) =>
+        'Hello ' + (args.who ?? 'World'),
+    },
+    thrower: {
+      type: GraphQLString,
+      resolve() {
+        throw new Error('Throws!');
+      },
+    },
+  },
+});
+
+export const TestSchema = new GraphQLSchema({
+  query: QueryRootType,
+  mutation: new GraphQLObjectType({
+    name: 'MutationRoot',
+    fields: {
+      writeTest: {
+        type: QueryRootType,
+        resolve: () => ({}),
+      },
+    },
+  }),
+});
